Copy TimingMate_v2 and updates in list views
I create it and use it to finish my tasks effectively. I set an allowed
time to finish for a task. Like I plan to use 3 hours to finish my
Deception Detection project. Every time I do the project, I set a
Timer. It is just like a smaller goal for today maybe 30 mins which
makes me focus on the task. PS: I just redo and redesign the app a few
days before. Some of the functionalities are not implemented and the UI
isn't good. The new version of the app is based on user testing for the
previous version. I plan to use another week to polish it.